Refactor amsuGetHelper
