+++ Source/Portal/CodeHelpers/amsuGetHelper.h
@@ -1,0 +1,10 @@
+﻿#pragma once
+
+#include "CoreMinimal.h"
+
+// TODO think about an upgrade and refactor - use templates
+
+struct amsuGetHelper
+{
+	static APlayerController* GetPlayerController(AActor* Actor);
+};
